+++ spec/unit/ogr/spatial_reference_mixins/exporters_spec.rb
@@ -23,18 +23,14 @@ RSpec.describe OGR::SpatialReference do
     context 'known projection' do
       subject { described_class.new_from_epsg(4322) }
 
-      it 'returns a populated array' do
-        pending 'Understanding how to make it work'
-        expect(subject.to_mapinfo).to be_an Array
-        expect(subject.to_mapinfo).to_not be_empty
+      it 'returns the string' do
+        expect(subject.to_mapinfo).to start_with 'Earth Projection'
       end
     end
 
     context 'empty SRS' do
-      it 'raises an OGR::NotEnoughData' do
-        pending 'Understanding how to make it work'
-
-        expect { subject.to_mapinfo }.to raise_exception OGR::NotEnoughData
+      it 'returns a "NonEarth Units" message' do
+        expect(subject.to_mapinfo).to start_with 'NonEarth Units'
       end
     end
   end
@@ -68,7 +64,8 @@ RSpec.describe OGR::SpatialReference do
       subject { described_class.new_from_epsg(4322) }
 
       it 'returns a PROJ4 String' do
-        expect(subject.to_proj4.strip).to eq('+proj=longlat +ellps=WGS72 +towgs84=0,0,4.5,0,0,0.554,0.2263 +no_defs')
+        expect(subject.to_proj4.strip)
+          .to eq('+proj=longlat +ellps=WGS72 +towgs84=0,0,4.5,0,0,0.554,0.2263 +no_defs')
       end
     end
 
@@ -99,26 +96,14 @@ RSpec.describe OGR::SpatialReference do
     context 'known projection' do
       subject { described_class.new_from_epsg(4322) }
 
-      it 'returns a well-known text String' do
-        expect(subject.to_pretty_wkt).to eq <<~WKT.strip
-          GEOGCS["WGS 72",
-              DATUM["WGS_1972",
-                  SPHEROID["WGS 72",6378135,298.26,
-                      AUTHORITY["EPSG","7043"]],
-                  TOWGS84[0,0,4.5,0,0,0.554,0.2263],
-                  AUTHORITY["EPSG","6322"]],
-              PRIMEM["Greenwich",0,
-                  AUTHORITY["EPSG","8901"]],
-              UNIT["degree",0.0174532925199433,
-                  AUTHORITY["EPSG","9122"]],
-              AUTHORITY["EPSG","4322"]]
-        WKT
+      it 'returns a formatted well-known text String' do
+        expect(subject.to_wkt).to start_with %(GEOGCS["WGS 72",DATUM[")
       end
     end
 
     context 'empty SRS' do
       it 'returns an empty String' do
-        expect(subject.to_pretty_wkt).to start_with 'GEOGCS["WGS 72",DATUM["'
+        expect(subject.to_pretty_wkt).to be_empty
       end
     end
   end
@@ -128,8 +113,7 @@ RSpec.describe OGR::SpatialReference do
       subject { described_class.new_from_epsg(4322) }
 
       it 'returns an XML string' do
-        pending 'Understanding how to make it work'
-        expect(subject.to_xml).to eq 'something'
+        expect(subject.to_xml).to start_with '<gml:GeographicCRS gml:id="ogrcrs1">'
       end
     end
 
@@ -138,5 +122,9 @@ RSpec.describe OGR::SpatialReference do
         expect { subject.to_xml }.to raise_exception OGR::UnsupportedSRS
       end
     end
+  end
+
+  describe '#to_gml' do
+    it { is_expected.to respond_to :to_gml }
   end
 end